pyproject.toml
[project]
name = "frozenplanetoid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "alex", email = "alex@pdp7.net" }
]
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "pyyaml>=6.0.2",
]

[project.scripts]
frozenplanetoid = "frozenplanetoid:main"
frozenplanetoid-opml-to-fp-yaml = "frozenplanetoid.utils:opml_to_yaml"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "ruff>=0.12.7",
]
